ddelta_apply: Fix warnings of different signedness
ddelta_apply.c: In function ‘copy_file’:
ddelta_apply.c:38:25: warning: comparison of integer expressions of different signedness: ‘long unsigned int’ and ‘off_t’ {aka ‘long int’} [-Wsign-compare]
   38 | #define MIN(x, y) (((x) < (y)) ? (x) : (y))
      |                         ^
ddelta_apply.c:196:27: note: in expansion of macro ‘MIN’
  196 |         uint32_t toread = MIN(sizeof(buf), end - start);
      |                           ^~~
ddelta_apply.c:38:40: warning: operand of ‘?:’ changes signedness from ‘off_t’ {aka ‘long int’} to ‘long unsigned int’ due to unsignedness of other operand [-Wsign-compare]
   38 | #define MIN(x, y) (((x) < (y)) ? (x) : (y))
      |                                        ^~~
ddelta_apply.c:196:27: note: in expansion of macro ‘MIN’
  196 |         uint32_t toread = MIN(sizeof(buf), end - start);
      |                           ^~~

+++ ddelta_apply.c
@@ -193,7 +193,7 @@ static int copy_file(const char *a, FILE *b, off_t start, off_t end, uint32_t *c
     }
 
     while (start < end && err >= 0) {
-        uint32_t toread = MIN(sizeof(buf), end - start);
+        size_t toread = MIN(sizeof(buf), (size_t)(end - start));
 
         if (fread(&buf, toread, 1, af) < 1) {
             ddelta_debug("copy_file failed.\n");
@@ -237,7 +237,7 @@ static int compute_crc32(FILE *a, off_t start, off_t end, uint32_t *crc)
     }
 
     while (start < end && err >= 0) {
-        uint32_t toread = MIN(sizeof(buf), end - start);
+        size_t toread = MIN(sizeof(buf), (size_t)(end - start));
 
         if (fread(&buf, toread, 1, a) < 1) {
             ddelta_debug("compute_crc32 failed.\n");
